Refactor location of test mocks
